# documentation/api/openapi.yaml
openapi: 3.0.3
info:
  title: BudgetApp API
  description: >
    Contratos iniciales para el MVP de BudgetApp (onboarding, presupuestos,
    transacciones, alertas e integraciones bancarias). Esta especificación sirve
    como base para generar SDKs/mocks y evolucionará conforme avancemos en la
    fase de Elaboración.
  version: 0.1.0
  contact:
    name: BudgetApp Team
servers:
  - url: https://api.budgetapp.local
    description: Sandbox local
  - url: https://api.staging.budgetapp.com
    description: Staging
...
paths:
...
  /api/transactions:
    post:
      tags: [Transactions]
      summary: Registra transacción manual.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionInput'
      responses:
        '201':
          description: Transacción creada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
    get:
      tags: [Transactions]
      summary: Lista transacciones filtradas por estado/origen.
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: source
          schema:
            type: string
            enum: [manual, bank]
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, posted]
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
        - in: query
          name: householdId
          schema:
            type: string
            nullable: true
          description: Filtra por hogar cuando se trata de un paquete familiar.
      responses:
        '200':
          description: Lista paginada.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Transaction'
                  nextCursor:
                    type: string
...
  /api/tickets:
    get:
      tags: [Tickets]
      summary: Lista los tickets del usuario (o todos si es admin/cs_agent).
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de tickets.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Ticket'
    post:
      tags: [Tickets]
      summary: Crea un ticket de soporte.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TicketInput'
      responses:
        '201':
          description: Ticket creado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ticket'
  /api/tickets/{id}:
    put:
      tags: [Tickets]
      summary: Actualiza estado/asignación o agrega comentarios.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [open, in_progress, resolved, closed]
                assignedTo:
                  type: string
                priority:
                  type: string
                  enum: [low, medium, high]
                comment:
                  type: string
      responses:
        '200':
          description: Ticket actualizado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ticket'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    TransactionInput:
      type: object
      properties:
        householdId:
          type: string
          nullable: true
          description: Identificador del hogar cuando la transacción afecta al presupuesto familiar.
        country:
          type: string
          minLength: 2
          maxLength: 3
        date:
          type: string
          format: date-time
        amount:
          type: number
        categoryId:
          type: string
        method:
          type: string
          enum: [cash, debit_card, credit_card, transfer]
        notes:
          type: string
        attachments:
          type: array
          items:
            type: string
            format: uri
      required: [country, date, amount, categoryId]
    Transaction:
      allOf:
        - $ref: '#/components/schemas/TransactionInput'
        - type: object
          properties:
            id:
              type: string
            status:
              type: string
              enum: [pending, posted]
            source:
              type: string
              enum: [manual, bank]
...
    TicketComment:
      type: object
      properties:
        id:
          type: string
        authorId:
          type: string
        message:
          type: string
        createdAt:
          type: string
          format: date-time
    Ticket:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        priority:
          type: string
          enum: [low, medium, high]
        status:
          type: string
          enum: [open, in_progress, resolved, closed]
        createdBy:
          type: string
        assignedTo:
          type: string
          nullable: true
        comments:
          type: array
          items:
            $ref: '#/components/schemas/TicketComment'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, title, description, status, priority]
    TicketInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        priority:
          type: string
          enum: [low, medium, high]
        assignedTo:
          type: string
      required: [title, description]
...